codified select track command
